+++ src/integrations/index.ts
@@ -5,3 +5,4 @@ export { nativeReleaseIntegration } from './release';
 export { capacitorRewriteFramesIntegration } from './rewriteframes';
 export { sdkInfoIntegration } from './sdkinfo';
 export { spotlightIntegration } from './spotlight';
+export { capacitorHttpIntegration } from './capacitorHttp';