+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ## 4.4.0
 
+### Features
+
+- Add automatic instrumentation for `CapacitorHttp` requests, including spans, breadcrumbs, and trace propagation using `tracePropagationTargets` and `propagateTraceparent`. ([#1387](https://github.com/getsentry/sentry-capacitor/pull/1387))
+- 
 ### Fixes
 
 - Pin Sentry Cocoa to 9.28.0 ([#1404](https://github.com/getsentry/sentry-capacitor/pull/1404))